avg enrollment goes from int to double

+++ src/tacos.mvc/Data/Course.cs
@@ -10,6 +10,6 @@ namespace tacos.data
         public string Name { get; set; }
 
         public double AverageSectionsPerCourse { get; set; }
-        public int AverageEnrollment { get; set; }
+        public double AverageEnrollment { get; set; }
     }
 }
