+++ src/tacos.mvc/Data/Request.cs
@@ -23,7 +23,7 @@ namespace tacos.data
 
         public double AverageSectionsPerCourse { get; set; }
 
-        public int AverageEnrollment { get; set; }
+        public double AverageEnrollment { get; set; }
 
         // calculated total of TAs, regardless of what is requested
         public double CalculatedTotal { get; set; }
